Speed up results storage function
Switch from appending dataframe rows to keeping rows as dictionaries until the very end. We may be able to speed this up further with numpy if needed but this should be a sufficient speedup for this stage. Dropped runtime from ~1 min per run to ~0.2 mins per run for 3 day sim run.