Adjust earliest start date in advanced settings
This had broken due to the default start date now being earlier than the earliest allowed start date. Rather than being the current date, this is now hardcoded to 1st Jan 2023 - we may need to consider whether this is a good option long term, and whether this is something that could/should be managed via the state file or some other constants file that is easier for people picking up this code for adaptation to adjust.